+++ src/api-response.ts
@@ -1,17 +1,31 @@
 import { z } from "zod";
-import { MimeDefinition, mimeJson } from "./mime";
 
-export type ApiResponse<A = z.ZodTypeAny> = {
-  schema: A;
-  mimeTypes: string[];
-};
+export interface ApiResponse<S extends z.ZodType> {
+  schema: S;
+  /**
+   * @default 200 for a positive response
+   * @default 400 for a negative response
+   * */
+  statusCode?: number;
+  /**
+   * @default "application/json"
+   * @override mimeTypes
+   * */
+  mimeType?: string;
+  /** @default [ "application/json" ] */
+  mimeTypes?: [string, ...string[]];
+}
 
-export const createApiResponse = <S extends z.ZodTypeAny>(
+/**
+ * @deprecated replace with just a schema, or { schema, mimeType } or { schema, mimeTypes } object
+ * @todo remove in v9
+ */
+export const createApiResponse = <S extends z.ZodType>(
   schema: S,
-  mimeTypes: MimeDefinition = mimeJson
-) => {
-  return {
-    schema,
-    mimeTypes: typeof mimeTypes === "string" ? [mimeTypes] : mimeTypes,
-  } as ApiResponse<S>;
-};
+  mimeDefinition?: string | [string, ...string[]]
+): ApiResponse<S> => ({
+  schema,
+  ...(typeof mimeDefinition === "string"
+    ? { mimeType: mimeDefinition }
+    : { mimeTypes: mimeDefinition }),
+});