+++ src/open-api.ts
@@ -97,12 +97,12 @@ export class OpenAPI extends OpenApiBuilder {
       const operation: OperationObject = {
         operationId: this.ensureUniqOperationId(path, method),
         responses: {
-          "200": depictResponse({
+          [endpoint.getPositiveStatusCode()]: depictResponse({
             ...commonParams,
             description: successfulResponseDescription,
             isPositive: true,
           }),
-          "400": depictResponse({
+          [endpoint.getNegativeStatusCode()]: depictResponse({
             ...commonParams,
             description: errorResponseDescription,
             isPositive: false,